feat: 검증된 GameManager 모델 기능 구현
