+++ README.md
@@ -54,9 +54,9 @@
 ---
 - Manager
   - 게임 매니저는 게임 규칙을 관리한다.
-    - [ ] 딜러와 플레이어는 게임이 시작되면 차례대로 한 장씩 2장의 카드를 받는다.
-    - [ ] 플레이어들은 `피니시` 상태가 아니라면 카드를 더 받을 수 있다.
-    - [ ] 플레이어들은 `액션` 상태라면 카드 추가 여부를 결정할 수 있다.
+    - [x] 딜러와 플레이어는 게임이 시작되면 차례대로 한 장씩 2장의 카드를 받는다.
+    - [x] 플레이어들은 `피니시` 상태가 아니라면 카드를 더 받을 수 있다.
+    - [x] 플레이어들은 `액션` 상태라면 카드 추가 여부를 결정할 수 있다.
 - Result
   - 결과는 게임 결과를 관리한다.
   - 결과는 참여자들을 가지고 있다.
