+++ src/test/kotlin/blackjack/model/GameManagerTest.kt
@@ -4,25 +4,6 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 
 class GameManagerTest {
-    class GameManger {
-        private val cardDeck = CardDeck()
-
-        fun setGame(participants: Participants) {
-            repeat(INIT_HAND_CARD_COUNT) {
-                participants.getParticipants().forEach { participant ->
-                    participant.draw(cardDeck.draw())
-                }
-            }
-        }
-
-        fun applyUserDrawDecision(participant: Participant) {
-            participant.draw(cardDeck.draw())
-        }
-
-        companion object {
-            const val INIT_HAND_CARD_COUNT: Int = 2
-        }
-    }
 
     @Test
     fun `블랙잭 게임 초기상태 설정 확인 테스트, (2장씩 갖고 있어야 함)`() {
